Add dbmate dependency for migrations, and create an initial migration
