+++ app/src/database/index.ts
@@ -1,0 +1,35 @@
+import fs from "fs";
+import os from "os";
+import path from "path";
+import Database from "better-sqlite3";
+
+let db: Database.Database | null = null;
+let databaseUrl: string | null = null;
+
+export const openDatabase = () => {
+  if (db) {
+    return db;
+  }
+
+  // Ensure the directory exists
+  const dbDir = path.join(os.homedir(), ".config", "SecureDrop");
+  if (!fs.existsSync(dbDir)) {
+    fs.mkdirSync(dbDir, { recursive: true });
+  }
+
+  // Create or open the SQLite database
+  const dbPath = path.join(dbDir, "db.sqlite");
+  db = new Database(dbPath, {});
+  db.pragma("journal_mode = WAL");
+
+  // Set the database URL for migrations
+  databaseUrl = `sqlite:${dbPath}`;
+  return db;
+};
+
+export const closeDatabase = () => {
+  if (db) {
+    db.close();
+    db = null;
+  }
+};
